+++ crates/full-node/sov-sequencer/tests/integration/main.rs
@@ -1,5 +1,6 @@
 mod preferred_end_to_end;
 mod preferred_with_reorgs;
+mod preferred_with_replicas;
 mod standard_sequencer;
 mod thin_sequencer;
 mod upgradability;